changed the uint32_t in the prototype of unpack_cell to size_t

+++ old/mimocat.h
@@ -41,5 +41,5 @@ typedef struct unpacked_cell {  /* how we store cells after unpacking them but
 } unpacked_cell_t;
 
 
-uint8_t *unpack_cell(uint8_t *, uint32_t , unpacked_cell_t *);
+uint8_t *unpack_cell(uint8_t *, size_t , unpacked_cell_t *);
 void pack_cell(cell_hdr_t *, uint8_t *, packed_cell_t *);
